Apply refactor from 6.4 branch
