+++ packages/block-editor/src/components/block-rename/is-empty-string.js
@@ -1,0 +1,3 @@
+export default function isEmptyString( testString ) {
+	return testString?.trim()?.length === 0;
+}
