+++ packages/block-editor/src/style.scss
@@ -15,6 +15,7 @@
 @import "./components/block-patterns-paging/style.scss";
 @import "./components/block-popover/style.scss";
 @import "./components/block-preview/style.scss";
+@import "./components/block-rename/style.scss";
 @import "./components/block-settings-menu/style.scss";
 @import "./components/block-styles/style.scss";
 @import "./components/block-switcher/style.scss";
@@ -56,7 +57,6 @@
 @import "./hooks/padding.scss";
 @import "./hooks/position.scss";
 @import "./hooks/typography.scss";
-@import "./hooks/block-rename-ui.scss";
 
 @import "./components/block-toolbar/style.scss";
 @import "./components/inserter/style.scss";
